+++ app/bible/versions.tsx
@@ -1,0 +1,16 @@
+import BibleVersionsScreen from '@/src/screens/BibleVersionsScreen';
+import { Stack } from 'expo-router';
+
+export default function BibleVersionsRoute() {
+  return (
+    <>
+      <Stack.Screen
+        options={{
+          title: 'Bible Versions',
+          headerShadowVisible: false,
+        }}
+      />
+      <BibleVersionsScreen />
+    </>
+  );
+}
